requirements.txt
aiosqlite==0.17.0
asyncpg==0.27.0
iso8601==1.1.0
pydantic==1.10.2
pypika-tortoise==0.1.6
pytz==2022.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
